requirements.txt
blinker==1.7.0
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
Flask==3.0.2
Flask-Cors==4.0.0
graphql-core==3.2.3
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
pymongo==4.6.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.16.0
strawberry-graphql==0.220.0
typing_extensions==4.10.0
Werkzeug==3.0.1
